// src/type.ts
export type itemType = 'image' | 'text' | 'shape' | 'stage' | 'group';

// props
export interface IaddItem {
  type: itemType;
  value: string;
}

export type LayerIdType = string | number;
// ...
export interface IcommonInfo {
  id: LayerIdType;
  type: itemType;
  isSelected?: boolean;
  handleInfo: (a: any) => void;
  handleSelected?: (id: number) => void;
  setShowTransformer: (s: boolean) => void;
  stageRef?: any;
  myRef?: any;
  trRef?: any;
  onRef?: (a: any) => void;
  banDrag?: boolean;
  isNew?: boolean;
  x: number;
  y: number;
  // w: number;
  // h: number;
  scaleX?: number;
  scaleY?: number;
  stageScale: number;
  fontSize?: number;
  mType?: number;
  elementName?: string;
  label?: string; // 元素层名称
  name?: string; //psd解析出来的图层名称
}

export interface IimageInfo extends IcommonInfo {
  type: 'image';
  value: string;
  trRef: any;
  crop: any;
  width?: number;
  height?: number;
  _isAdaptStage?: number;
  _isProportionalScaling?: number;
  _isChangedCrop?: boolean;
}

interface IshapeCommon {
  id: LayerIdType;
}

interface RectProps extends Konva.RectConfig {}
interface CircleProps extends Konva.CircleConfig {}
interface ArcProps extends Konva.ArcConfig {}
interface StarProps extends Konva.StarConfig {}
interface ArrowProps extends Konva.ArrowConfig {}
interface EllipseProps extends Konva.EllipseConfig {}

export type ShapeType =
  | 'rect'
  | 'circle'
  | 'arc'
  | 'star'
  | 'arrow'
  | 'ellipse';

type ShapePropsMap = {
  rect: RectProps;
  circle: CircleProps;
  arc: ArcProps;
  star: StarProps;
  arrow: ArrowProps;
  ellipse: EllipseProps;
};
export type IShapeInfo = ShapePropsMap[ShapeType];
// ...
export interface IgroupInfo extends IcommonInfo {
  type: 'group';
  elements: Array<Iinfo>;
}

export interface ItextInfo extends IcommonInfo {
  type: 'text';
  value: string;
  color?: 'string';
  width?: number;
  height?: number;
  fontStyle?: string;
  stroke?: string;
  strokeWidth?: number;
}

export interface IFunc {
  exportToImage: (
    a: string,
    opt?: { scale?: number; quality?: number; fileType?: string }
  ) => void;
  exportToBASE64: () => Promise<string>;
  exportToFile: (format: string, filename: string) => File | undefined;
  withdraw: () => void;
  redo: () => void;
  canvasScale: (a: number) => void;
  deleteItem: () => void;
  copyItem: () => void;
  getInfo: () => any;
  moveLayerLevel: (i: number) => void;
  moveLayer: (direction: string, delta: number) => void;
  clearSelected: () => void;
  setSelectedIndex: (id: LayerIdType) => void;
  toogleLock: (id: LayerIdType) => void;
  toggleMultiSelected: (state: boolean) => void;
  madeGroup: (layers: any) => void;
  divideGroup: (groupId: string) => void;
  changeLayerInfoById: (id: LayerIdType, item: object) => void;
  // getSelectedInfo: () => Iinfo | Array<Iinfo>;
}

export type Iinfo = IimageInfo | ItextInfo | IShapeInfo | IgroupInfo;